# kubepuppy-app/kubepuppy/web/model_role_binding.go
/*
 * KubePuppy API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type RoleBinding struct {

	RoleBindingId string `json:"roleBindingId"`

	Kind string `json:"kind"`

	Name string `json:"name"`

	// For kind RoleBinding, the role binding namespace. Not applicable for kind ClusterRoleBinding.
	Namespace string `json:"namespace,omitempty"`
}