# kubepuppy-app/kubepuppy/web/model_subject_details.go
/*
 * KubePuppy API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type SubjectDetails struct {

	RoleBindings []SubjectDetailsRoleBinding `json:"roleBindings"`

	Id string `json:"id"`

	Kind string `json:"kind"`

	Name string `json:"name"`

	Namespace string `json:"namespace"`
}