# tmp/go/model_certificate copy.go
/*
 * KubePuppy API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Certificate struct {

	// X.509 certificate data
	Source string `json:"source"`

	// X.509 subject (in string format)
	SubjectAsText string `json:"subjectAsText,omitempty"`

	SubjectCommonName string `json:"subjectCommonName,omitempty"`

	SubjectOrganization string `json:"subjectOrganization,omitempty"`

	// X.509 issuer (in string format)
	Issuer string `json:"issuer,omitempty"`

	// Whether certificate is signed by cluster CA (and so can be used for K8S RBAC)
	IsAuthorised bool `json:"isAuthorised"`
}